ChatBot-Assistant-main/requirements.txt
langchain==0.1.14
langchain_openai
langchain_core
langchain_community
langchain-google-genai
langserve
streamlit==1.35.0
python-dotenv==1.0.1
tiktoken==0.7.0
faiss-cpu==1.8.0
chromadb==0.4.24
unstructured<0.12
beautifulsoup4==4.12.3
requests==2.31.0
nltk==3.8.1
sentence-transformers==2.6.1
redis==5.0.4
google-search-results==2.4.2
transformers==4.43.2
accelerate==0.31.0
joblib
pandas
scikit-learn
spacy